moved code to new file in xds package
